requirements.txt
python-dotenv
streamlit
Pillow
google-generativeai
IPython
langchain
langchain-google-genai
pandas
numpy
seaborn
boto3
nltk
textblob